fix: refactor network organization
